+++ hydromt/cli/main.py
@@ -16,7 +16,8 @@ from pydantic import ValidationError
 
 from hydromt.data_catalog import DataCatalog
 from hydromt.validators.data_catalog import DataCatalogValidator
-from hydromt.validators.model_config import HydromtModelStep
+from hydromt.validators.model_config import HydromtModelSetup
+from hydromt.validators.region import validate_region
 
 from .. import __version__, log
 from ..models import MODELS
@@ -332,39 +333,51 @@ def update(
 
 
 @main.command(
-    short_help="Validate config files are correct",
+    short_help="Validate config / data catalog / region",
 )
-@click.argument(
-    "MODEL",
+@click.option(
+    "-m",
+    "--model",
     type=str,
+    default=None,
+    help="Model name, e.g. wflow, sfincs, etc. to validate config file.",
 )
 @opt_config
 @data_opt
-@deltares_data_opt
 @quiet_opt
 @verbose_opt
+@region_opt
 @click.pass_context
 def check(
     ctx,
     model,
     config,
     data,
-    dd,
+    region: Optional[Dict[Any, Any]],
     quiet: int,
     verbose: int,
 ):
-    """Verify that provided data catalog files are in the correct format.
+    """
+    Verify that provided data catalog and config files are in the correct format.
+
+    Additionnaly region bbox and geom can also be validated.
 
     Example usage:
     --------------
 
-    hydromt check grid -d /path/to/data_catalog.yml -i /path/to/model_config.yml
+    Check data catalog file:
+    hydromt check -d /path/to/data_catalog.yml -v
+
+    Check data catalog and grid_model config file:
+    hydromt check -m grid_model -d /path/to/data_catalog.yml -i /path/to/model_config.yml -v
+
+    With region:
+    hydromt check -m grid_model -d /path/to/data_catalog.yml -i /path/to/model_config.yml -r '{'bbox': [-1,-1,1,1]}' -v
 
     """  # noqa: E501
     # logger
     log_level = max(10, 30 - 10 * (verbose - quiet))
     logger = log.setuplog("check", join(".", "hydromt.log"), log_level=log_level)
-    logger.info(f"Output dir: {export_dest_path}")
     try:
         all_exceptions = []
         for cat_path in data:
@@ -376,24 +389,36 @@ def check(
                 all_exceptions.append(e)
                 logger.info("Catalog has errors")
 
-        mod = MODELS.load(model)
-        try:
-            config_dict = cli_utils.parse_config(config)
-            logger.info(f"Validating config at {config}")
+        if region:
+            logger.info(f"Validating region {region}")
+            try:
+                validate_region(region)
+                logger.info("Region is valid!")
 
-            HydromtModelStep.from_dict(config_dict, model=mod)
-            logger.info("Model config valid!")
+            except (ValidationError, ValueError, NotImplementedError) as e:
+                logger.info("region has errors")
+                all_exceptions.append(e)
 
-        except (ValidationError, ValueError) as e:
-            logger.info("Model has errors")
-            all_exceptions.append(e)
+        if config:
+            mod = MODELS.load(model)
+            logger.info(f"Validating for model {model} of type {type(mod).__name__}")
+            try:
+                config_dict = cli_utils.parse_config(config)
+                logger.info(f"Validating config at {config}")
+
+                HydromtModelSetup.from_dict(config_dict, model=mod)
+                logger.info("Model config valid!")
+
+            except (ValidationError, ValueError) as e:
+                logger.info("Model has errors")
+                all_exceptions.append(e)
 
         if len(all_exceptions) > 0:
-            raise Exception(all_exceptions)
+            raise ValueError(all_exceptions)
 
     except Exception as e:
         logger.exception(e)  # catch and log errors
-        raise
+        raise e
     finally:
         for handler in logger.handlers[:]:
             handler.close()