+++ hydromt/models/model_api.py
@@ -22,6 +22,8 @@ from geopandas.testing import assert_geodataframe_equal
 from pyproj import CRS
 from shapely.geometry import box
 
+from hydromt import __version__
+from hydromt._compat import Distribution
 from hydromt.typing import DeferedFileClose, XArrayDict
 
 from .. import config, log, workflows
@@ -96,10 +98,17 @@ class Model(object, metaclass=ABCMeta):
         from . import MODELS  # avoid circular import
 
         self.logger = logger
-        dist, version = "unknown", "NA"
+        dist_name, version = "unknown", "NA"
         if self._NAME in MODELS:
             ep = MODELS[self._NAME]
-            dist, version = ep.distro.name, ep.distro.version
+            dist: Distribution | None = ep.dist
+            if dist:
+                dist_name = dist.name
+                version = dist.version
+            else:
+                # insert hydromt defaults
+                dist_name = "hydromt"
+                version = __version__
 
         # link to data
         self.data_catalog = DataCatalog(
@@ -130,7 +139,9 @@ class Model(object, metaclass=ABCMeta):
         # model paths
         self._config_fn = self._CONF if config_fn is None else config_fn
         self.set_root(root, mode)  # also creates hydromt.log file
-        self.logger.info(f"Initializing {self._NAME} model from {dist} (v{version}).")
+        self.logger.info(
+            f"Initializing {self._NAME} model from {dist_name} (v{version})."
+        )
 
     @_classproperty
     def api(cls) -> Dict:
@@ -1209,6 +1220,10 @@ class Model(object, metaclass=ABCMeta):
             )
         if name in self.geoms:  # trigger init / read
             self.logger.warning(f"Replacing geom: {name}")
+        if hasattr(self, "crs"):
+            # Verify if a geom is set to model crs and if not sets geom to model crs
+            if self.crs and self.crs != geom.crs:
+                geom.to_crs(self.crs.to_epsg(), inplace=True)
         self._geoms[name] = geom
 
     def read_geoms(self, fn: str = "geoms/*.geojson", **kwargs) -> None:
@@ -1232,7 +1247,9 @@ class Model(object, metaclass=ABCMeta):
             self.logger.debug(f"Reading model file {name}.")
             self.set_geoms(gpd.read_file(fn, **kwargs), name=name)
 
-    def write_geoms(self, fn: str = "geoms/{name}.geojson", **kwargs) -> None:
+    def write_geoms(
+        self, fn: str = "geoms/{name}.geojson", to_wgs84: bool = False, **kwargs
+    ) -> None:
         r"""Write model geometries to a vector file (by default GeoJSON) at <root>/<fn>.
 
         key-word arguments are passed to :py:meth:`geopandas.GeoDataFrame.to_file`
@@ -1242,6 +1259,8 @@ class Model(object, metaclass=ABCMeta):
         fn : str, optional
             filename relative to model root and should contain a {name} placeholder,
             by default 'geoms/{name}.geojson'
+        to_wgs84: bool, optional
+            Option to enforce writing GeoJSONs with WGS84(EPSG:4326) coordinates.
         \**kwargs:
             Additional keyword arguments that are passed to the
             `geopandas.to_file` function.
@@ -1250,8 +1269,6 @@ class Model(object, metaclass=ABCMeta):
             self.logger.debug("No geoms data found, skip writing.")
             return
         self._assert_write_mode()
-        if "driver" not in kwargs:
-            kwargs.update(driver="GeoJSON")  # default
         for name, gdf in self.geoms.items():
             if not isinstance(gdf, (gpd.GeoDataFrame, gpd.GeoSeries)) or len(gdf) == 0:
                 self.logger.warning(
@@ -1262,6 +1279,11 @@ class Model(object, metaclass=ABCMeta):
             _fn = join(self.root, fn.format(name=name))
             if not isdir(dirname(_fn)):
                 os.makedirs(dirname(_fn))
+            if to_wgs84 and (
+                kwargs.get("driver") == "GeoJSON"
+                or str(fn).lower().endswith(".geojson")
+            ):
+                gdf = gdf.to_crs(4326)
             gdf.to_file(_fn, **kwargs)
 
     @property