+++ docs/changelog.rst
@@ -15,6 +15,8 @@ Added
 -----
 - Export CLI now also accepts time tuples (#660)
 - New stats.skills VE and RSR (#666)
+- Check CLI command can now validate bbox and geom regions (#664)
+- Added GADM 4.1 as FlatGeoBuff files to deltares_data catalog (#686)
 
 
 Changed
@@ -24,9 +26,14 @@ Changed
 
 Fixed
 -----
+- Removed deprecated entrypoints library. (#676)
 - Bug in `raster.set_crs` if input_crs is of type CRS. (#659)
 - Export CLI now actually parses provided geoms. (#660)
 - Bug in stats.skills for computation of pbias and MSE / RMSE. (#666)
+- `Model.write_geoms` ow has an option to write GeoJSON coordinates in WGS84 if specified (#510)
+- Fix bug with lazy spatial_ref coordinate (#682)
+- Bug in gis_utils.meridian_offset. (#692)
+
 
 Deprecated
 ----------